Refactor dice roll and result calculation logic
This commit optimizes the dice roll and result evaluation functions in main.py. The `get_rolls` function is simplified using list comprehension and renamed to `get_roll` for clarity, as single dice roll functionality is better represented by this name. The logic of result evaluation function `get_result` is rewritten to cope with unpacking the list instead of comparing each item in a list which makes the code considerably cleaner and easier to understand. It also changes a few string outputs to allow for bank/PvP rule interpretation later.