Update string output for Trips results
This commit updates the string output for result in 'TRIPS' cases in main.py. The string output was formerly in title-case and has been updated to all uppercase to conform to conventions of other similar in-game notifications. This change facilitates the result readability and alignment with other game result scenarios when mid-game analytics are presented to the player.