Refactor game mechanics for improved result handling
The game mechanics have been significantly refactored to improve handling of roll-offs and ties between players. Previously, the game would arbitrarily declare a winner based on the initial roll, this has been updated to properly account for ties. Players who roll the same highest score will now enter into a 'roll-off' for the win. An additional message is displayed notifying all players of the roll-off. This refactoring will make the game more fair and exciting as all players now have an equal opportunity to compete for the win in case of a tie. It also sets the foundation for further improvements such as automating the roll-off process in subsequent game rounds.