Updated game mechanics, main, and player AI to support play order
This diff updates the game logic to accommodate a play order neccessary for 'BANK' games. Before there was no order in which players played which is fine for 'PvP' games. In 'BANK' games, however, the banker must roll first.

Functions 'roll_all_players', 'game_round_pvp', and 'determine_winner' are now equipped to handle ordering. A set of new functions 'determine_play_order', 'game_round_bank', and 'banker_reset' are introduced to support this feature.

The Player class is updated to include banker attribute, shaping the player to hold the information about being a banker. But it might be moved to the round itself.