Updated text outputs to improve user understanding
The changes were made to address ambiguity in our text output. In game_mechanics.py, the print statement was revised for wagering to ensure it was clear the total pot was the sum of wagers. In main.py, instructions for continuing after play order was determined were clearer and more intuitive.